chore: rename resolve-project-id
